Refactor testpotato command into modular utilities
Extract testpotato command logic from the monolithic command file into focused, reusable modules in src/mahoji/lib/testPotato/. This improves code organization, maintainability, and testability by separating concerns into distinct utility functions for gear, spawning, wiping, slayer tasks, monster KC, and other account management features.
