+++ src/mahoji/commands/testpotato.ts
@@ -1,343 +1,32 @@
-import { EmbedBuilder, userMention } from '@oldschoolgg/discord';
-import { noOp, stringMatches, Time } from '@oldschoolgg/toolkit';
-import { Bank, convertLVLtoXP, ItemGroups, Items, itemID, MAX_INT_JAVA, resolveItems } from 'oldschooljs';
+import { userMention } from '@oldschoolgg/discord';
+import { Items, itemID } from 'oldschooljs';
 
-import { xp_gains_skill_enum } from '@/prisma/main.js';
-import {
-	bitfieldCanUserManipulate,
-	changeBitFieldForUser,
-	getBitFieldData,
-	listBitFields
-} from '@/lib/bitFieldUtils.js';
-import { allStashUnitsFlat, allStashUnitTiers } from '@/lib/clues/stashUnits.js';
 import { CombatAchievements } from '@/lib/combat_achievements/combatAchievements.js';
 import { BitFieldData, globalConfig } from '@/lib/constants.js';
-import { spiritAnglerOutfit } from '@/lib/data/CollectionsExport.js';
-import { COXMaxMageGear, COXMaxMeleeGear, COXMaxRangeGear } from '@/lib/data/cox.js';
-import { leaguesCreatables } from '@/lib/data/creatables/leagueCreatables.js';
-import { Eatables } from '@/lib/data/eatables.js';
-import { TOBMaxMageGear, TOBMaxMeleeGear, TOBMaxRangeGear } from '@/lib/data/tob.js';
 import { effectiveMonsters } from '@/lib/minions/data/killableMonsters/index.js';
-import potions from '@/lib/minions/data/potions.js';
-import { MAX_QP, quests } from '@/lib/minions/data/quests.js';
-import { allOpenables } from '@/lib/openables.js';
+import { MAX_QP } from '@/lib/minions/data/quests.js';
 import { Minigames } from '@/lib/settings/minigames.js';
-import { Farming } from '@/lib/skilling/skills/farming/index.js';
-import {
-	farmingPatchNames,
-	getFarmingKeyFromName,
-	userGrowingProgressStr
-} from '@/lib/skilling/skills/farming/utils/farmingHelpers.js';
-import { getFarmingInfoFromUser } from '@/lib/skilling/skills/farming/utils/getFarmingInfo.js';
+import { farmingPatchNames } from '@/lib/skilling/skills/farming/utils/farmingHelpers.js';
 import { Skills } from '@/lib/skilling/skills/index.js';
 import { slayerMasterChoices } from '@/lib/slayer/constants.js';
-import { slayerMasters } from '@/lib/slayer/slayerMasters.js';
-import { SlayerRewardsShop } from '@/lib/slayer/slayerUnlocks.js';
 import { allSlayerMonsters } from '@/lib/slayer/tasks/index.js';
-import { Gear } from '@/lib/structures/Gear.js';
-import type { SafeUserUpdateInput } from '@/lib/user/update.js';
-import { parseStringBank } from '@/lib/util/parseStringBank.js';
 import { fetchBingosThatUserIsInvolvedIn } from '@/mahoji/commands/bingo.js';
-import { gearViewCommand } from '@/mahoji/lib/abstracted_commands/gearCommands.js';
-import { getPOH } from '@/mahoji/lib/abstracted_commands/pohCommand.js';
-import { shades, shadesLogs } from '@/mahoji/lib/abstracted_commands/shadesOfMortonCommand.js';
-import { allUsableItems } from '@/mahoji/lib/abstracted_commands/useCommand.js';
 import { BingoManager } from '@/mahoji/lib/bingo/BingoManager.js';
+import { handleTestPotatoBingoTools } from '@/mahoji/lib/testPotato/bingoTools.js';
+import { handleTestPotatoBitfield } from '@/mahoji/lib/testPotato/bitfield.js';
+import { handleTestPotatoConfirmation } from '@/mahoji/lib/testPotato/confirmation.js';
+import { handleTestPotatoForceGrow } from '@/mahoji/lib/testPotato/forceGrow.js';
+import { handleTestPotatoGear } from '@/mahoji/lib/testPotato/gear.js';
+import { getMaxUserValues, giveMaxStats, maxUser } from '@/mahoji/lib/testPotato/max.js';
+import { gearPresets, spawnPresets } from '@/mahoji/lib/testPotato/presets.js';
+import { thingsToReset, thingsToWipe } from '@/mahoji/lib/testPotato/resetOptions.js';
+import { handleTestPotatoSetMonsterKC } from '@/mahoji/lib/testPotato/setMonsterKC.js';
+import { handleTestPotatoSetSlayerTask } from '@/mahoji/lib/testPotato/setSlayerTask.js';
+import { setMinigameKC, setXP } from '@/mahoji/lib/testPotato/setters.js';
+import { handleTestPotatoSpawn } from '@/mahoji/lib/testPotato/spawn.js';
+import { handleTestPotatoWipe } from '@/mahoji/lib/testPotato/wipe.js';
 
-export function getMaxUserValues(): SafeUserUpdateInput {
-	const updates: SafeUserUpdateInput = {};
-	for (const skill of Object.values(xp_gains_skill_enum)) {
-		updates[`skills_${skill}`] = convertLVLtoXP(99);
-	}
-	return {
-		minion_hasBought: true,
-		QP: MAX_QP,
-		slayer_points: 50_000,
-		nmz_points: 50_000,
-		volcanic_mine_points: 500_000,
-		carpenter_points: 5_000_000,
-		zeal_tokens: 500_000,
-		lms_points: 500_000,
-		slayer_unlocks: SlayerRewardsShop.map(i => i.id),
-		...updates
-	};
-}
-
-export async function giveMaxStats(user: MUser) {
-	return user.update(getMaxUserValues());
-}
-
-const coloMelee = new Gear();
-for (const gear of Items.resolveItems([
-	'Torva full helm',
-	'Infernal cape',
-	'Amulet of blood fury',
-	'Torva platebody',
-	'Torva platelegs',
-	'Primordial boots',
-	'Ultor ring',
-	'Scythe of vitur'
-])) {
-	coloMelee.equip(Items.getOrThrow(gear));
-}
-
-const coloRange = new Gear();
-for (const gear of Items.resolveItems([
-	"Dizana's quiver",
-	'Masori mask (f)',
-	'Necklace of anguish',
-	'Masori body (f)',
-	'Masori chaps (f)',
-	'Pegasian boots',
-	'Venator ring',
-	'Dragon arrow',
-	'Twisted bow'
-])) {
-	coloRange.equip(Items.getOrThrow(gear));
-}
-
-const gearPresets = [
-	{
-		name: 'Cox',
-		melee: COXMaxMeleeGear,
-		mage: COXMaxMageGear,
-		range: COXMaxRangeGear
-	},
-	{
-		name: 'ToB',
-		melee: TOBMaxMeleeGear,
-		mage: TOBMaxMageGear,
-		range: TOBMaxRangeGear
-	},
-	{
-		name: 'Colosseum',
-		melee: coloMelee,
-		range: coloRange,
-		mage: coloRange
-	}
-];
-
-const thingsToReset = [
-	{
-		name: 'Everything/All',
-		run: async (user: MUser) => {
-			await prisma.slayerTask.deleteMany({ where: { user_id: user.id } }).catch(noOp);
-			await prisma.activity.deleteMany({ where: { user_id: BigInt(user.id) } }).catch(noOp);
-			await prisma.commandUsage.deleteMany({ where: { user_id: BigInt(user.id) } }).catch(noOp);
-			await prisma.gearPreset.deleteMany({ where: { user_id: user.id } }).catch(noOp);
-			await prisma.giveaway.deleteMany({ where: { user_id: user.id } }).catch(noOp);
-			await prisma.lastManStandingGame.deleteMany({ where: { user_id: BigInt(user.id) } }).catch(noOp);
-			await prisma.minigame.deleteMany({ where: { user_id: user.id } }).catch(noOp);
-			await prisma.newUser.deleteMany({ where: { id: user.id } }).catch(noOp);
-			await prisma.playerOwnedHouse.deleteMany({ where: { user_id: user.id } }).catch(noOp);
-			await prisma.user.deleteMany({ where: { id: user.id } }).catch(noOp);
-			return 'Reset all your data.';
-		}
-	},
-	{
-		name: 'Bank',
-		run: async (user: MUser) => {
-			await prisma.user.update({
-				where: {
-					id: user.id
-				},
-				data: {
-					bank: {}
-				}
-			});
-			return 'Reset your bank.';
-		}
-	}
-];
-
-async function setMinigameKC(user: MUser, _minigame: string, kc: number) {
-	if (_minigame.toLowerCase() === 'all') {
-		await prisma.minigame.update({
-			where: {
-				user_id: user.id
-			},
-			data: Object.fromEntries(Minigames.map(game => [game.column, kc]))
-		});
-		await user.statsUpdate({
-			tithe_farms_completed: kc
-		});
-		return `Set all ${Minigames.length} minigame KCs to ${kc}.`;
-	}
-	const minigame = Minigames.find(m => m.column === _minigame.toLowerCase());
-	if (!minigame) return 'No kc set because invalid minigame.';
-	await prisma.minigame.update({
-		where: {
-			user_id: user.id
-		},
-		data: {
-			[minigame.column]: kc
-		}
-	});
-	if (minigame.column === 'tithe_farm') {
-		await user.statsUpdate({
-			tithe_farms_completed: kc
-		});
-	}
-	return `Set your ${minigame.name} KC to ${kc}.`;
-}
-
-async function setXP(user: MUser, skillName: string, xp: number) {
-	if (skillName === 'all') {
-		await user.update(
-			Object.fromEntries(Object.values(xp_gains_skill_enum).map(enumSkill => [`skills_${enumSkill}`, xp]))
-		);
-		return `Set all ${Object.values(xp_gains_skill_enum).length} skills to ${xp} XP.`;
-	}
-	const skill = Object.values(Skills).find(c => c.id === skillName);
-	if (!skill) return 'No xp set because invalid skill.';
-	await user.update({
-		[`skills_${skill.id}`]: xp
-	});
-	return `Set ${skill.name} XP to ${xp}.`;
-}
-const openablesBank = new Bank();
-for (const i of allOpenables.values()) {
-	openablesBank.add(i.id, 100);
-}
-
-const equippablesBank = new Bank();
-for (const i of Items.filter(i => Boolean(i.equipment) && Boolean(i.equipable)).values()) {
-	if (ItemGroups.allUnobtainableItems.includes(i.id)) continue;
-	equippablesBank.add(i.id);
-}
-
-const farmingPreset = new Bank();
-for (const plant of Farming.Plants) {
-	farmingPreset.add(plant.inputItems.clone().multiply(100));
-	if (plant.protectionPayment) {
-		farmingPreset.add(plant.protectionPayment.clone().multiply(100));
-	}
-}
-farmingPreset.add('Ultracompost', 10_000);
-const usables = new Bank();
-for (const usable of allUsableItems) usables.add(usable, 100);
-
-const leaguesPreset = new Bank();
-for (const a of leaguesCreatables) leaguesPreset.add(a.outputItems);
-
-const allStashUnitItems = new Bank();
-for (const unit of allStashUnitsFlat) {
-	for (const i of [unit.items].flat(2)) {
-		allStashUnitItems.add(i);
-	}
-}
-for (const tier of allStashUnitTiers) {
-	allStashUnitItems.add(tier.cost.clone().multiply(tier.units.length));
-}
-
-const potionsPreset = new Bank();
-for (const potion of potions) {
-	for (const actualPotion of potion.items) {
-		potionsPreset.addItem(actualPotion, 100_000);
-	}
-}
-
-const foodPreset = new Bank();
-for (const food of Eatables.map(food => food.id)) {
-	foodPreset.addItem(food, 100_000);
-}
-
-const anglerOutfit = resolveItems(['Angler hat', 'Angler top', 'Angler waders', 'Angler boots']);
-const fishingPreset = new Bank()
-	.add('Fish sack barrel')
-	.add('Fish barrel')
-	.add("Rada's blessing 4")
-	.add('Shark lure', MAX_INT_JAVA)
-	.add('Feather', MAX_INT_JAVA)
-	.add('Fishing bait', MAX_INT_JAVA)
-	.add('Raw karambwanji', MAX_INT_JAVA)
-	.add('Fishing bait', MAX_INT_JAVA)
-	.add('Sandworms', MAX_INT_JAVA)
-	.add('Spirit flakes', MAX_INT_JAVA)
-	.add('Crystal shard', MAX_INT_JAVA)
-	.add('Stripy feather', MAX_INT_JAVA)
-	.add('Pearl fly fishing rod')
-	.add('Pearl fishing rod')
-	.add('Fly fishing rod')
-	.add('Fishing rod')
-	.add('Oily fishing rod')
-	.add('Crystal harpoon')
-	.add('Infernal harpoon')
-	.add('Dark fishing bait', MAX_INT_JAVA);
-for (const anglerPiece of [...spiritAnglerOutfit, ...anglerOutfit]) {
-	fishingPreset.add(anglerPiece);
-}
-
-const runePreset = new Bank()
-	.add('Air rune', MAX_INT_JAVA)
-	.add('Mind rune', MAX_INT_JAVA)
-	.add('Water rune', MAX_INT_JAVA)
-	.add('Earth rune', MAX_INT_JAVA)
-	.add('Fire rune', MAX_INT_JAVA)
-	.add('Body rune', MAX_INT_JAVA)
-	.add('Cosmic rune', MAX_INT_JAVA)
-	.add('Chaos rune', MAX_INT_JAVA)
-	.add('Nature rune', MAX_INT_JAVA)
-	.add('Law rune', MAX_INT_JAVA)
-	.add('Death rune', MAX_INT_JAVA)
-	.add('Astral rune', MAX_INT_JAVA)
-	.add('Blood rune', MAX_INT_JAVA)
-	.add('Soul rune', MAX_INT_JAVA)
-	.add('Dust rune', MAX_INT_JAVA)
-	.add('Lava rune', MAX_INT_JAVA)
-	.add('Mist rune', MAX_INT_JAVA)
-	.add('Mud rune', MAX_INT_JAVA)
-	.add('Smoke rune', MAX_INT_JAVA)
-	.add('Steam rune', MAX_INT_JAVA);
-
-const shadesPreset = new Bank().add('Olive oil(4)', 100_000).add('Sacred oil(4)', 100_000);
-for (const log of shadesLogs) {
-	shadesPreset.add(log.normalLog.id, 100_000);
-	shadesPreset.add(log.oiledLog.id, 100_000);
-}
-for (const shade of shades) {
-	shadesPreset.add(shade.item.id, 100_000);
-	if (shade.lowMetalKeys) {
-		for (const key of shade.lowMetalKeys.items) {
-			if (!shadesPreset.has(key)) shadesPreset.add(key, 100_000);
-		}
-	}
-	if (shade.highMetalKeys) {
-		for (const key of shade.highMetalKeys.items) {
-			if (!shadesPreset.has(key)) shadesPreset.add(key, 100_000);
-		}
-	}
-}
-for (const coffin of ['Bronze coffin', 'Steel coffin', 'Black coffin', 'Silver coffin', 'Gold coffin']) {
-	shadesPreset.add(coffin, 1);
-}
-
-const spawnPresets = [
-	['fishing', fishingPreset],
-	['openables', openablesBank],
-	['random', new Bank()],
-	['equippables', equippablesBank],
-	['farming', farmingPreset],
-	['usables', usables],
-	['leagues', leaguesPreset],
-	['stashunits', allStashUnitItems],
-	['potions', potionsPreset],
-	['food', foodPreset],
-	['runes', runePreset],
-	['shades', shadesPreset]
-] as const;
-
-const thingsToWipe = [
-	'bank',
-	'combat_achievements',
-	'cl',
-	'quests',
-	'buypayout',
-	'kc',
-	'cooldowns',
-	'birdhouses',
-	'giveaways'
-] as const;
+export { getMaxUserValues, giveMaxStats };
 
 export const testPotatoCommand = globalConfig.isProduction
 	? null
@@ -751,87 +440,14 @@ export const testPotatoCommand = globalConfig.isProduction
 					};
 				}
 				if (options.confirmation) {
-					const ephemeral = options.confirmation.ephemeral ?? false;
-					const users = [user.id];
-					if (options.confirmation.other_person) users.push(options.confirmation.other_person.user.id);
-					if (options.confirmation.another_person) users.push(options.confirmation.another_person.user.id);
-					if (ephemeral && users.length > 1) {
-						return 'You cannot have multiple people confirm on an ephemeral message.';
-					}
-					await interaction.confirmation({
-						content: `This is a normal confirmation. Users who must confirm: ${users.map(i => `<@${i}>`).join(', ')}`,
-						users,
-						// @ts-expect-error
-						ephemeral
-					});
-					return interaction.makePaginatedMessage({
-						ephemeral: true,
-						pages: [
-							() => ({
-								embeds: [
-									new EmbedBuilder()
-										.setTitle(`Page 1`)
-										.setImage(`https://cdn.oldschool.gg/monkey/${rng.randInt(1, 39)}.webp`)
-								]
-							}),
-							() => ({
-								embeds: [
-									new EmbedBuilder()
-										.setTitle(`Page 2`)
-										.setImage(`https://cdn.oldschool.gg/monkey/${rng.randInt(1, 39)}.webp`)
-								]
-							}),
-							() => ({
-								embeds: [
-									new EmbedBuilder()
-										.setTitle(`Page 3`)
-										.setImage(`https://cdn.oldschool.gg/monkey/${rng.randInt(1, 39)}.webp`)
-								]
-							}),
-							() => ({
-								embeds: [
-									new EmbedBuilder()
-										.setTitle(`Page 4`)
-										.setImage(`https://cdn.oldschool.gg/monkey/${rng.randInt(1, 39)}.webp`)
-								]
-							})
-						]
-					});
+					return handleTestPotatoConfirmation(user, interaction, rng, options.confirmation);
 				}
 
 				if (options.bitfield) {
-					if (!options.bitfield.add && !options.bitfield.remove) {
-						return 'you must choose a valid bitfield from either add or remove';
-					}
-					const aor = () => Boolean(options.bitfield!.add);
-
-					const bitInput = options.bitfield.add ?? options.bitfield.remove!;
-					const bit = getBitFieldData(bitInput);
-					const action: 'add' | 'remove' = aor() ? 'add' : 'remove';
-					if (!bit) return listBitFields(user);
-					const canManipulate = bitfieldCanUserManipulate({ user, bit });
-					if (canManipulate !== true) return canManipulate;
-					return changeBitFieldForUser(user, bit.bit, action);
+					return handleTestPotatoBitfield(user, options.bitfield);
 				}
 				if (options.bingo_tools) {
-					if (options.bingo_tools.start_bingo) {
-						const bingo = await prisma.bingo.findFirst({
-							where: {
-								id: Number(options.bingo_tools.start_bingo),
-								creator_id: user.id
-							}
-						});
-						if (!bingo) return 'Invalid bingo.';
-						await prisma.bingo.update({
-							where: {
-								id: bingo.id
-							},
-							data: {
-								start_date: new Date()
-							}
-						});
-						return 'Your bingo start date has been set to this moment, so it has just started.';
-					}
+					return handleTestPotatoBingoTools(user, options.bingo_tools);
 				}
 
 				if (options.set) {
@@ -859,203 +475,13 @@ export const testPotatoCommand = globalConfig.isProduction
 					return `You now ${!current ? 'ARE' : 'ARE NOT'} an ironman.`;
 				}
 				if (options.wipe) {
-					const { thing } = options.wipe;
-					if (thing === 'birdhouses') {
-						await user.updateBirdhouseData({
-							lastPlaced: null,
-							birdhousePlaced: false,
-							birdhouseTime: 0
-						});
-						return 'Reset your birdhouses.';
-					}
-					if (thing === 'giveaways') {
-						await prisma.giveaway.deleteMany({
-							where: {
-								user_id: user.id
-							}
-						});
-						return 'Wiped all your giveaways (no refunds given).';
-					}
-					if (thing === 'cooldowns') {
-						await user.update({
-							gambling_lockout_expiry: null
-						});
-						await prisma.userStats.upsert({
-							where: {
-								user_id: BigInt(user.id)
-							},
-							update: {
-								last_daily_timestamp: Date.now() - Time.Day,
-								last_tears_of_guthix_timestamp: Date.now() - Time.Day * 2
-							},
-							create: {
-								user_id: BigInt(user.id)
-							}
-						});
-						return 'Reset all your daily/TOG cooldowns, gambling lockout.';
-					}
-					if (thing === 'kc') {
-						await user.statsUpdate({
-							monster_scores: {}
-						});
-						return 'Reset all your KCs.';
-					}
-					if (thing === 'buypayout') {
-						await prisma.botItemSell.deleteMany({
-							where: {
-								user_id: user.id
-							}
-						});
-						return 'Deleted all your buy payout records, so you have no tax rate accumulated.';
-					}
-					if (thing === 'bank') {
-						await user.update({
-							// @ts-expect-error
-							bank: {}
-						});
-						return 'Reset your bank.';
-					}
-					if (thing === 'cl') {
-						await user.update({
-							// @ts-expect-error
-							collectionLogBank: {},
-							temp_cl: {}
-						});
-						await prisma.userStats.update({
-							where: {
-								user_id: BigInt(user.id)
-							},
-							data: {
-								cl_array: [],
-								cl_array_length: 0
-							}
-						});
-						return 'Reset your collection log.';
-					}
-					if (thing === 'combat_achievements') {
-						await user.update({
-							completed_ca_task_ids: []
-						});
-						return 'Reset your combat achievements.';
-					}
-					if (thing === 'quests') {
-						await prisma.user.update({
-							where: {
-								id: user.id
-							},
-							data: {
-								finished_quest_ids: [],
-								collectionLogBank: {}
-							}
-						});
-						return `Your QP, and completed quests, have been reset. You can set your QP to a certain number using ${globalClient.mentionCommand(
-							'testpotato',
-							'set'
-						)}.`;
-					}
-					return 'Invalid thing to reset.';
+					return handleTestPotatoWipe(user, options.wipe.thing);
 				}
 				if (options.max) {
-					await getPOH(user.id);
-					await prisma.playerOwnedHouse.update({
-						where: {
-							user_id: user.id
-						},
-						data: {
-							pool: 29_241
-						}
-					});
-					await roboChimpClient.user.upsert({
-						where: {
-							id: BigInt(user.id)
-						},
-						create: {
-							id: BigInt(user.id),
-							leagues_points_balance_osb: 25_000
-						},
-						update: {
-							leagues_points_balance_osb: {
-								increment: 25_000
-							}
-						}
-					});
-					await user.addItemsToBank({
-						items: new Bank()
-							.add('Rune pouch')
-							.add('Blood rune', 100_000_000)
-							.add('Death rune', 100_000_000)
-							.add('Blood rune', 100_000_000)
-							.add('Water rune', 100_000_000)
-							.add('Saradomin brew(4)', 100_000_000)
-							.add('Super restore(4)', 100_000_000)
-							.add('Stamina potion(4)', 100_000_000)
-							.add('Super combat potion(4)', 100_000_000)
-							.add('Cooked karambwan', 100_000_000)
-							.add('Ranging potion(4)', 100_000_000)
-							.add('Coins', 100_000_000)
-							.add('Shark', 100_000_000)
-							.add('Vial of blood', 100_000_000)
-							.add('Rune pouch')
-							.add('Zamorakian spear')
-							.add('Dragon warhammer')
-							.add('Bandos godsword')
-							.add('Toxic blowpipe')
-							.add(runePreset)
-							.add(foodPreset)
-							.add(potionsPreset)
-							.add(usables)
-					});
-
-					await user.updateGear([
-						{ setup: 'melee', gear: COXMaxMeleeGear.raw() },
-						{ setup: 'range', gear: COXMaxRangeGear.raw() },
-						{ setup: 'mage', gear: COXMaxMageGear.raw() }
-					]);
-
-					await user.rawUpdate({
-						data: {
-							GP: 5000000000,
-							slayer_points: 100000,
-							tentacle_charges: 10000,
-							sang_charges: 10000,
-							trident_charges: 10000,
-							serp_helm_charges: 10000,
-							blood_fury_charges: 10000,
-							tum_shadow_charges: 10000,
-							blood_essence_charges: 10000,
-							ash_sanctifier_charges: 10000,
-							celestial_ring_charges: 10000,
-							scythe_of_vitur_charges: 10000,
-							venator_bow_charges: 10000,
-							blowpipe: {
-								scales: 100000,
-								dartQuantity: 100000,
-								dartID: itemID('Dragon dart')
-							},
-							finished_quest_ids: quests.map(q => q.id)
-						}
-					});
-					await giveMaxStats(user);
-					return 'Fully maxed your account, stocked your bank, charged all chargeable items.';
+					return maxUser(user);
 				}
 				if (options.gear) {
-					const gear = gearPresets.find(i => stringMatches(i.name, options.gear?.thing))!;
-
-					for (const type of ['melee', 'range', 'mage'] as const) {
-						const currentGear = gear[type];
-						const ammo = currentGear.get('ammo');
-						if (ammo && Items.getItem(ammo.item)?.stackable) {
-							currentGear.set('ammo', { ...ammo, quantity: 10_000 });
-						}
-					}
-
-					await user.updateGear([
-						{ setup: 'melee', gear: gear.melee.raw() },
-						{ setup: 'range', gear: gear.range.raw() },
-						{ setup: 'mage', gear: gear.mage.raw() }
-					]);
-
-					return gearViewCommand(user, 'all', false);
+					return handleTestPotatoGear(user, options.gear.thing);
 				}
 				if (options.reset) {
 					const resettable = thingsToReset.find(i => i.name === options.reset?.thing);
@@ -1069,162 +495,19 @@ export const testPotatoCommand = globalConfig.isProduction
 					return setXP(user, options.setxp.skill, options.setxp.xp);
 				}
 				if (options.spawn) {
-					const { preset, collectionlog, item, items } = options.spawn;
-					const bankToGive = new Bank();
-					if (preset) {
-						const actualPreset = spawnPresets.find(i => i[0] === preset);
-						if (!actualPreset) return 'Invalid preset';
-						let b = actualPreset[1];
-						if (actualPreset[0] === 'random') {
-							b = new Bank();
-							for (let i = 0; i < 1000; i++) {
-								b.add(Items.random().id);
-							}
-						}
-						bankToGive.add(b);
-					}
-					if (item) {
-						try {
-							bankToGive.add(Items.getOrThrow(item).id);
-						} catch (err) {
-							return err as string;
-						}
-					}
-					if (items) {
-						for (const [i, qty] of parseStringBank(items, undefined, true)) {
-							bankToGive.add(i.id, qty || 1);
-						}
-					}
-
-					await user.addItemsToBank({ items: bankToGive, collectionLog: Boolean(collectionlog) });
-					return `Spawned: ${bankToGive.toString().slice(0, 1800)}.`;
+					return handleTestPotatoSpawn(user, options.spawn);
 				}
 
 				if (options.setmonsterkc) {
-					if (options.setmonsterkc.monster.toLowerCase() === 'all') {
-						const kc = options.setmonsterkc.kc ?? 1;
-						const stats = await user.fetchStats();
-						await user.statsUpdate({
-							monster_scores: {
-								...(stats.monster_scores as Record<string, number>),
-								...Object.fromEntries(effectiveMonsters.map(mon => [mon.id, kc]))
-							}
-						});
-						return `Set all ${effectiveMonsters.length} monster KCs to ${kc}.`;
-					}
-					const monster = effectiveMonsters.find(m =>
-						stringMatches(m.name, options.setmonsterkc?.monster ?? '')
-					);
-					if (!monster) return 'Invalid monster';
-					const stats = await user.fetchStats();
-					await user.statsUpdate({
-						monster_scores: {
-							...(stats.monster_scores as Record<string, unknown>),
-							[monster.id]: options.setmonsterkc?.kc ?? 1
-						}
-					});
-					return `Set your ${monster.name} KC to ${options.setmonsterkc.kc ?? 1}.`;
+					return handleTestPotatoSetMonsterKC(user, options.setmonsterkc);
 				}
 
 				if (options.forcegrow) {
-					if (options.forcegrow.patch_name === 'birdhouses') {
-						const birdhouseData = user.fetchBirdhouseData();
-						await user.updateBirdhouseData({
-							...birdhouseData,
-							birdhouseTime: Date.now() - Time.Month
-						});
-						return 'Your birdhouses have been forced to be fully grown.';
-					}
-					const farmingDetails = await getFarmingInfoFromUser(user);
-					const { patch_name } = options.forcegrow;
-					const patchesToGrow =
-						patch_name === 'all'
-							? farmingDetails.patchesDetailed.filter(patch => patch.plant)
-							: farmingDetails.patchesDetailed.filter(
-									patch => patch.patchName === patch_name && patch.plant
-								);
-					if (patchesToGrow.length === 0) {
-						return patch_name === 'all'
-							? 'You have nothing planted in any patches.'
-							: 'You have nothing planted there.';
-					}
-					const now = Date.now();
-					const updates: SafeUserUpdateInput = Object.fromEntries(
-						patchesToGrow.map(patch => [
-							getFarmingKeyFromName(patch.patchName),
-							{
-								...farmingDetails.patches[patch.patchName],
-								plantTime: now - Time.Month
-							}
-						])
-					);
-
-					await user.update(updates);
-					return userGrowingProgressStr((await getFarmingInfoFromUser(user)).patchesDetailed, user);
+					return handleTestPotatoForceGrow(user, options.forcegrow.patch_name);
 				}
 
 				if (options.setslayertask) {
-					const usersTask = await user.fetchSlayerInfo();
-
-					const { monster, master } = options.setslayertask;
-
-					const selectedMonster = allSlayerMonsters.find(m => stringMatches(m.name, monster));
-					const selectedMaster = slayerMasters.find(
-						sm => stringMatches(master, sm.name) || sm.aliases.some(alias => stringMatches(master, alias))
-					);
-					if (!selectedMaster || !selectedMonster) return 'Invalid slayer master or monster.';
-
-					// Set quantity to 50 if user doesn't assign a quantity
-					const quantity = options.setslayertask?.quantity ?? 50;
-
-					const assignedTask = selectedMaster.tasks.find(m => m.monster.id === selectedMonster.id)!;
-
-					if (!assignedTask) {
-						const possibleMasters = slayerMasters
-							.filter(m => m.tasks.some(t => t.monster.id === selectedMonster?.id))
-							.map(m => m.name);
-
-						const suggestion =
-							possibleMasters.length > 0
-								? ` (${possibleMasters.join(', ')} can${possibleMasters.length > 1 ? '' : ' also'} assign this monster.)`
-								: '';
-
-						return `${selectedMaster.name} cannot assign ${selectedMonster.name}.${suggestion}`;
-					}
-
-					// Update an existing slayer task for the user
-					if (usersTask.currentTask?.id) {
-						await prisma.slayerTask.update({
-							where: {
-								id: usersTask.currentTask?.id
-							},
-							data: {
-								quantity,
-								quantity_remaining: quantity,
-								slayer_master_id: selectedMaster.id,
-								monster_id: selectedMonster.id,
-								skipped: false
-							}
-						});
-					} else {
-						// Create a new slayer task for the user
-						await prisma.slayerTask.create({
-							data: {
-								user_id: user.id,
-								quantity,
-								quantity_remaining: quantity,
-								slayer_master_id: selectedMaster.id,
-								monster_id: selectedMonster.id,
-								skipped: false
-							}
-						});
-					}
-
-					await user.update({
-						slayer_last_task: selectedMonster.id
-					});
-
-					return `You set your slayer task to ${selectedMonster.name} using ${selectedMaster.name}.`;
+					return handleTestPotatoSetSlayerTask(user, options.setslayertask);
 				}
 
 				return 'Nothin!';
